ROVER-244 Remove dead code and tidy up
