--- src/command/dev/legacy/mod.rs
@@ -1,26 +1,0 @@
-#[cfg(feature = "composition-js")]
-mod compose;
-
-#[cfg(feature = "composition-js")]
-mod do_dev;
-
-#[cfg(feature = "composition-js")]
-mod introspect;
-
-#[cfg(feature = "composition-js")]
-mod protocol;
-
-#[cfg(feature = "composition-js")]
-mod router;
-
-#[cfg(feature = "composition-js")]
-mod schema;
-
-#[cfg(feature = "composition-js")]
-mod netstat;
-
-#[cfg(not(feature = "composition-js"))]
-mod no_dev;
-
-#[cfg(feature = "composition-js")]
-mod watcher;
