--- src/command/dev/legacy/router/mod.rs
@@ -1,7 +1,0 @@
-mod command;
-mod config;
-mod runner;
-
-pub use command::{BackgroundTask, BackgroundTaskLog};
-pub use config::RouterConfigHandler;
-pub use runner::RouterRunner;
