+++ src/command/dev/router/watchers/router_config.rs
@@ -1,12 +1,11 @@
 use futures::StreamExt;
 use tap::TapFallible;
 
+use crate::command::dev::FileWatcher;
 use crate::{
-    command::dev::next::router::{config::RouterConfig, hot_reload::RouterUpdateEvent},
+    command::dev::router::{config::RouterConfig, hot_reload::RouterUpdateEvent},
     subtask::SubtaskHandleUnit,
 };
-
-use crate::command::dev::next::FileWatcher;
 
 /// Watches for router config changes
 pub struct RouterConfigWatcher {
